// backend/src/model/formdata.ts
export interface IFormData {
	university: string;
	sport: string;
	sportType: string;
	team: string;
	role: string;
	number: string;
	prefix: string;
	name: string;
	shirt: string;
	pants: string;
	phone: string;
	studentId: string;
	allergies: string;
	email: string;
}

export interface IUrlEntry {
	university: string;
	sport: string;
	sportType: string;
	team: string;
	role: string;
	number: string;
	prefix: string;
	name: string;
	shirt: string;
	pants: string;
	phone: string;
	studentId: string;
	allergies: string;
	email: string;
}

export type regisTracking = {
	id: number;
	sport: string;
	sex: string;
	type: string;
	institute: string;
	status: boolean;
};